test(lib.rs): load configuration test
Created a test to test the load_configuration function. The test creates a temp directory and sets
the KOMOREBI_CONFIG_HOME to the temp directory. The test will create a mock powershell config and
call the load_configuration function and check to see if the config file was loaded successfully.
The test will then clean up the temp directory.