test(monitor.rs): test moving a container to a nonexistent workspace
Created a test that test moving a container to a workspace that doesn't exist. The test check that
when a container is moved to a workspace that doesn't exist, the workspace is created and that the
workspace contains the container. It also checks that there n number of workspace available where n
is the largest workspace number.