test(workspace.rs): test removing a non existent window
Created a test that tests removing a window that doesn't exist. The test creates a container with a
window and then attempts to remove a window that doesn't exist. The test will check the result to
ensure that the result returned an error and that we still have the expected number of windows.